requirements.txt
tensorflow
Flask
numpy
Pillow
matplotlib
seaborn
scikit-learn